+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Unreleased
 
 * Removal: `Scalingo::Client#agora_fr1` had been removed since the region no longer exists.
+* New: Add `addons#authenticate` endpoint
+* New API: databases API
+* New API: backups API
 
 ## 3.1.0