+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 * Removal: `Scalingo::Client#agora_fr1` had been removed since the region no longer exists.
 * New: Add `addons#authenticate` endpoint
-* New API: databases API
-* New API: backups API
+* New API: database API
+* New API: backup API
 
 ## 3.1.0