+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 4.0.0
+
 * Removal: `Scalingo::Client#agora_fr1` had been removed since the region no longer exists.
 * New: Add `addons#authenticate!` endpoint
 * New API: database API